third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_url_file=webkit02-expected.txt

This is a testharness.js-based test.
[FAIL] <select><input>
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|     <input>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <input>"
Harness: the test ran to completion.
